// src/app/modules/financeiro/fluxocaixa/enums/financeiro-fluxocaixa-enums.model.ts
/**********************************************************************************************
Code generated with MKL Plug-in version: 40.2.5
Code generated at time stamp: 2019-12-31T10:28:19.070
Copyright: Kerubin - logokoch@gmail.com

WARNING: DO NOT CHANGE THIS CODE BECAUSE THE CHANGES WILL BE LOST IN THE NEXT CODE GENERATION.
***********************************************************************************************/

import { SelectItem } from 'primeng/api';

export interface TipoFonteMovimento extends SelectItem {
	
  label?: string;
  value: any;
  
}

export interface FormaPagamento extends SelectItem {
	
  label?: string;
  value: any;
  
}

export interface TipoReceitaDespesa extends SelectItem {
	
  label?: string;
  value: any;
  
}

export interface TipoContaBancaria extends SelectItem {
	
  label?: string;
  value: any;
  
}

export interface CaixaDiarioSituacao extends SelectItem {
	
  label?: string;
  value: any;
  
}

export interface TipoLancamentoFinanceiro extends SelectItem {
	
  label?: string;
  value: any;
  
}

export interface TipoPlanoContaFinanceiro extends SelectItem {
	
  label?: string;
  value: any;
  
}

export interface TipoPessoa extends SelectItem {
	
  label?: string;
  value: any;
  
}